\name{getEnrichmentTable-dispatch}
\alias{getEnrichmentTable}
\title{
Method dispatch page for getEnrichmentTable
}
\description{
Method dispatch page for \code{getEnrichmentTable}.
}
\section{Dispatch}{
\code{getEnrichmentTable} can be dispatched on following classes:

\itemize{
\item \code{\link{getEnrichmentTable,GreatObject-method}}, \code{\link{GreatObject-class}} class method
\item \code{\link{getEnrichmentTable,GreatJob-method}}, \code{\link{GreatJob-class}} class method
}
}
\examples{
# no example
NULL


}
