\name{getEnrichmentTables-dispatch}
\alias{getEnrichmentTables}
\title{
Method dispatch page for getEnrichmentTables
}
\description{
Method dispatch page for \code{getEnrichmentTables}.
}
\section{Dispatch}{
\code{getEnrichmentTables} can be dispatched on following classes:

\itemize{
\item \code{\link{getEnrichmentTables,GreatJob-method}}, \code{\link{GreatJob-class}} class method
\item \code{\link{getEnrichmentTables,GreatObject-method}}, \code{\link{GreatObject-class}} class method
}
}
\examples{
# no example
NULL


}
