\name{getEnrichmentTables-GreatObject-method}
\alias{getEnrichmentTables,GreatObject-method}
\title{
Get enrichment table
}
\description{
Get enrichment table
}
\usage{
\S4method{getEnrichmentTables}{GreatObject}(object, ...)
}
\arguments{

  \item{object}{A \code{\link{GreatObject-class}} object returned by \code{\link{great}}.}
  \item{...}{All passed to \code{\link{getEnrichmentTable,GreatObject-method}}.}

}
\details{
Please use \code{\link{getEnrichmentTable,GreatObject-method}} directly.
}
\examples{
# There is no example
NULL

}
