\name{getGenesFromGencode}
\alias{getGenesFromGencode}
\title{
Get Gencode genes
}
\description{
Get Gencode genes
}
\usage{
getGenesFromGencode(version)
}
\arguments{

  \item{version}{Gencode version, e.g. v19 for human, vM21 for mouse.}

}
\details{
Only the protein coding genes.
}
\value{
A \code{\link[GenomicRanges:GRanges-class]{GRanges}} object.
}
\examples{
# There is no example
NULL

}
