\name{getGREATDefaultTSS}
\alias{getGREATDefaultTSS}
\title{
Get built-in TSS from GREAT
}
\description{
Get built-in TSS from GREAT
}
\usage{
getGREATDefaultTSS(genome)
}
\arguments{

  \item{genome}{Only support "hg19", "hg38", "mm10", "mm9". Files are downloaded from \url{https://great-help.atlassian.net/wiki/spaces/GREAT/pages/655445/Genes} .}

}
\value{
A \code{\link[GenomicRanges:GRanges-class]{GRanges}} object.
}
\examples{
# There is no example
NULL

}
