\name{getKEGGGenome}
\alias{getKEGGGenome}
\title{
Get the corresponding assembly id for a kegg organism
}
\description{
Get the corresponding assembly id for a kegg organism
}
\usage{
getKEGGGenome(organism)
}
\arguments{

  \item{organism}{The organism code on KEGG.}

}
\value{
The Refseq access ID for the genome.
}
\examples{
# There is no example
NULL

}
