\name{getKEGGPathways}
\alias{getKEGGPathways}
\title{
Get KEGG pathway gene sets
}
\description{
Get KEGG pathway gene sets
}
\usage{
getKEGGPathways(organism, as_table = FALSE)
}
\arguments{

  \item{organism}{The organism code on KEGG.}
  \item{as_table}{Whether to return the gene sets as a two-column table.}

}
\value{
A list of a data frame, depends on the value of \code{as_table}.
}
\examples{
# There is no example
NULL

}
