\name{getRegionGeneAssociations-dispatch}
\alias{getRegionGeneAssociations}
\title{
Method dispatch page for getRegionGeneAssociations
}
\description{
Method dispatch page for \code{getRegionGeneAssociations}.
}
\section{Dispatch}{
\code{getRegionGeneAssociations} can be dispatched on following classes:

\itemize{
\item \code{\link{getRegionGeneAssociations,GreatJob-method}}, \code{\link{GreatJob-class}} class method
\item \code{\link{getRegionGeneAssociations,GreatObject-method}}, \code{\link{GreatObject-class}} class method
}
}
\examples{
# no example
NULL


}
