\name{getTSS}
\alias{getTSS}
\title{
Get the internally used TSS
}
\description{
Get the internally used TSS
}
\usage{
getTSS(tss_source, biomart_dataset = NULL)
}
\arguments{

  \item{tss_source}{The same format as in \code{\link{great}}.}
  \item{biomart_dataset}{The same format as in \code{\link{great}}.}

}
\value{
A \code{\link[GenomicRanges:GRanges-class]{GRanges}} object.
}
\examples{
# There is no example
NULL

}
