\name{great_opt}
\alias{great_opt}
\title{
Global parameters for rGREAT
}
\description{
Global parameters for rGREAT
}
\usage{
great_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE, ADD = FALSE)
}
\arguments{

  \item{...}{Arguments for the parameters, see "details" section}
  \item{RESET}{Reset to default values.}
  \item{READ.ONLY}{Please ignore.}
  \item{LOCAL}{Pllease ignore.}
  \item{ADD}{Please ignore.}

}
\details{
There are following parameters:

\describe{
  \item{\code{verbose}}{Whether to show messages.}
}
}
\examples{
great_opt
}
