\name{GreatJob-class}
\docType{class}
\alias{GreatJob-class}
\alias{GreatJob}
\title{
Class to store and retrieve GREAT results  


}
\description{
Class to store and retrieve GREAT results  


}
\details{
After submitting request to GREAT server, the generated results will be  available on GREAT server for some time. The \code{GreatJob} class is defined  to store parameters that user has set and result tables what are retrieved  from GREAT server.  


}
\section{contructor}{
\code{\link{submitGreatJob}} is used to generate a \code{GreatJob} instance.  


}
\author{
Zuguang gu <z.gu@dkfz.de>  


}
\examples{
# please refer to page of `submitGreatJob`
}
