\name{GreatJob}
\alias{GreatJob}
\title{
Constructor method for GreatJob class
}
\description{
Constructor method for GreatJob class
}
\usage{
GreatJob(...)
}
\arguments{

  \item{...}{arguments.}

}
\details{
There is no public constructor method for the \code{\link{GreatJob-class}}.
}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
