\name{GreatObject}
\alias{GreatObject}
\title{
Constructor method for GreatObject class
}
\description{
Constructor method for GreatObject class
}
\usage{
GreatObject(...)
}
\arguments{

  \item{...}{arguments.}

}
\details{
There are following methods that can be applied on \code{\link{GreatObject-class}} object:

\itemize{
  \item \code{\link{getEnrichmentTable,GreatObject-method}} to retrieve the result table. 
  \item \code{\link{getRegionGeneAssociations,GreatObject-method}} to get the associations between input regions and genes.
  \item \code{\link{plotRegionGeneAssociations,GreatObject-method}} to plot the associations bewteen input regions and genes.
  \item \code{\link{shinyReport,GreatObject-method}} to view the results by a shiny application.
}
}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
