\name{plotRegionGeneAssociations-dispatch}
\alias{plotRegionGeneAssociations}
\title{
Method dispatch page for plotRegionGeneAssociations
}
\description{
Method dispatch page for \code{plotRegionGeneAssociations}.
}
\section{Dispatch}{
\code{plotRegionGeneAssociations} can be dispatched on following classes:

\itemize{
\item \code{\link{plotRegionGeneAssociations,GreatObject-method}}, \code{\link{GreatObject-class}} class method
\item \code{\link{plotRegionGeneAssociations,GreatJob-method}}, \code{\link{GreatJob-class}} class method
}
}
\examples{
# no example
NULL


}
