\name{plotVolcano-dispatch}
\alias{plotVolcano}
\title{
Method dispatch page for plotVolcano
}
\description{
Method dispatch page for \code{plotVolcano}.
}
\section{Dispatch}{
\code{plotVolcano} can be dispatched on following classes:

\itemize{
\item \code{\link{plotVolcano,GreatJob-method}}, \code{\link{GreatJob-class}} class method
\item \code{\link{plotVolcano,GreatObject-method}}, \code{\link{GreatObject-class}} class method
}
}
\examples{
# no example
NULL


}
