\name{shinyReport-dispatch}
\alias{shinyReport}
\title{
Method dispatch page for shinyReport
}
\description{
Method dispatch page for \code{shinyReport}.
}
\section{Dispatch}{
\code{shinyReport} can be dispatched on following classes:

\itemize{
\item \code{\link{shinyReport,GreatJob-method}}, \code{\link{GreatJob-class}} class method
\item \code{\link{shinyReport,GreatObject-method}}, \code{\link{GreatObject-class}} class method
}
}
\examples{
# no example
NULL


}
