\name{shinyReport-GreatObject-method}
\alias{shinyReport,GreatObject-method}
\title{
Shiny app on the GreatObject object
}
\description{
Shiny app on the GreatObject object
}
\usage{
\S4method{shinyReport}{GreatObject}(object)
}
\arguments{

  \item{object}{The \code{GreatObject} object returned by \code{\link{great}}.}

}
\value{
A shiny app object.
}
\examples{
if(FALSE) {
# pseudo code
obj = great(...)
shinyReport(obj)
}
}
