\docType{class}
\name{GREAT_Job-class}
\alias{GREAT_Job}
\alias{GREAT_Job-class}
\title{Reference class to store GREAT results}
\description{
Reference class to store GREAT results
}
\section{Fields}{

\describe{
\item{\code{id}}{job id for current GREAT job}
\item{\code{parameters}}{Parameters that user sets}
\item{\code{enrichment_tables}}{Tables for enrichment in different ontologies}
\item{\code{association_tables}}{Tables for association between genomic regions and genes}

}}