\name{plotRegionGeneAssociationGraphs-GreatJob-method}
\alias{plotRegionGeneAssociationGraphs,GreatJob-method}
\alias{plotRegionGeneAssociationGraphs}
\title{
Plot region-gene associations
}
\description{
Plot region-gene associations
}
\usage{
\S4method{plotRegionGeneAssociationGraphs}{GreatJob}(object, ...)
}
\arguments{

  \item{object}{A \code{\link{GreatJob-class}} object returned by \code{\link{submitGreatJob}}.}
  \item{...}{All passed to \code{\link{plotRegionGeneAssociations,GreatJob-method}}.}

}
\details{
This function will be removed in the future, please use \code{\link{plotRegionGeneAssociations,GreatJob-method}} instead.
}
\examples{
# There is no example
NULL

}
